Move the Delete rule button into the rule redit page
Move the delete rule button into the rule edit page. This is where we
are going to put the disable rule switch too, as there were becoming too
many actions on the rules list page.
